refactor(sdk): fold local_transport into remote_zip module
